requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.6
psycopg2-binary==2.9.7